// client/src/api/model/initiativeSlotType.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type InitiativeSlotType = typeof InitiativeSlotType[keyof typeof InitiativeSlotType];


export const InitiativeSlotType = {
  Player: 'Player',
  NPC: 'NPC',
} as const;